move renderer to scratch2 package
