+++ browser.js
@@ -2,7 +2,7 @@ function makeCanvas() {
   return document.createElement("canvas")
 }
 
-var scratchblocks = (window.scratchblocks = module.exports = require("./lib/")(
+var scratchblocks = (window.scratchblocks = module.exports = require("./scratch2/")(
   window,
   makeCanvas
 ))
