+++ syntax/index.js
@@ -6,11 +6,17 @@ const {
   Input,
   Block,
   Comment,
+  Glow,
   Script,
   Document,
 } = require("./model.js")
 
+const { allLanguages, loadLanguages } = require("./blocks.js")
+
 module.exports = {
+  allLanguages,
+  loadLanguages,
+
   parse,
   fromJSON: Document.fromJSON,
 
@@ -19,6 +25,7 @@ module.exports = {
   Input,
   Block,
   Comment,
+  Glow,
   Script,
   Document,
 }
